Move initialization to another file
