+++ kmeans_functions.py
@@ -1,33 +1,15 @@
 import numpy as np
 import scipy.spatial.distance as ssdist
+import initialization as init
 
 
 def initialization(n_clusters, initialize, x):
     # this method initializes the first set of labels
-    datasize = x.shape[0]
     if initialize == 'random':
-        np.random.seed(45)
-        indices = np.random.choice(datasize, n_clusters, replace=False)
-        centroids = x[indices]
+        centroids = init.random_init(x, n_clusters)
     elif initialize == 'k-means++':
-        first_cent = np.random.choice(datasize, 1)
-        centroids = x[first_cent]
-        k = 1
-        while k < n_clusters:
-            new_centre = _do_for_kpp(centroids, x)
-            centroids = np.append(centroids, new_centre, axis=0)
-            k += 1
+        centroids = init.kpp_init(x, n_clusters)
     return centroids
-
-
-def _do_for_kpp(centroids, x):
-    datasize = x.shape[0]
-    dist = fix_the_distance(x, centroids, 'correlation')
-    p_dist = np.amin(dist, axis=1)
-    all_dist = np.sum(p_dist)
-    probability = [(p_dist[i]) / all_dist for i in range(datasize)]
-    new_centre = x[np.random.choice(datasize, 1, p=probability)]
-    return new_centre
 
 
 def labeling(centroids, x):
